moved plot_handler to plotting module and updated imports
